Reduce split packages in o.e.tests
Split packages cause a lot of troubles and this is test harness code so
it's better to live without them.
