+++ eclipse.platform.releng/bundles/org.eclipse.test.performance/META-INF/MANIFEST.MF
@@ -6,18 +6,18 @@ Bundle-Version: 3.20.300.qualifier
 Bundle-Activator: org.eclipse.test.internal.performance.PerformanceTestPlugin
 Bundle-Vendor: %Plugin.providerName
 Bundle-Localization: plugin
+Import-Package: org.junit.jupiter.api,
+ org.junit.platform.suite.api,
+ org.eclipse.test
 Export-Package: org.eclipse.perfmsr.core,
- org.eclipse.test,
- org.eclipse.test.internal;x-internal:=true,
  org.eclipse.test.internal.performance,
  org.eclipse.test.internal.performance.data,
  org.eclipse.test.internal.performance.db,
  org.eclipse.test.internal.performance.eval,
  org.eclipse.test.internal.performance.tests,
  org.eclipse.test.performance
-Import-Package: org.junit.jupiter.api,
- org.junit.platform.suite.api
 Require-Bundle: org.eclipse.core.runtime,
+ org.eclipse.test;visibility:=reexport,
  org.junit
 Bundle-ActivationPolicy: lazy
 Bundle-ClassPath: .
