+++ eclipse.platform.releng/bundles/org.eclipse.test/META-INF/MANIFEST.MF
@@ -10,7 +10,8 @@ Require-Bundle: org.apache.ant,
  org.eclipse.ui;bundle-version="[3.112.0,4.0.0)",
  org.eclipse.core.runtime,
  org.eclipse.ui.ide.application,
- org.eclipse.equinox.app
+ org.eclipse.equinox.app,
+ org.junit
 Import-Package: org.junit.jupiter.api,
  org.junit.platform.engine,
  org.junit.platform.engine.discovery,
