+++ eclipse.platform.releng/bundles/org.eclipse.test/src/org/eclipse/test/TracingSuite.java
